+++ app/src/main/java/com/example/candyshop/DetailsScreen.kt
@@ -5,7 +5,12 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
@@ -14,6 +19,9 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.example.candyshop.ui.theme.ContentItem
+import com.example.candyshop.ui.theme.content
+
 
 fun getCandyDetailsByName(name: String?): ContentItem? {
     var result: ContentItem? = null
@@ -23,6 +31,17 @@ fun getCandyDetailsByName(name: String?): ContentItem? {
         }
     }
     return result
+}
+
+@Composable
+fun SimpleFilledTextFieldSample() {
+    var text by remember { mutableStateOf("Hello") }
+
+    TextField(
+        value = text,
+        onValueChange = { text = it },
+        label = { Text("Label") }
+    )
 }
 
 @Composable
